sig verification

+++ scripts/ff-tb-updater.sh
@@ -1,25 +1,33 @@
 #!/bin/bash
 #
 # Author: Morfik
-# Version: 0.0.1
-# Date: 2015.03.22
+# Version: 0.1.0
+# Date: 2015.04.12
 #
 # See: https://wiki.mozilla.org/Software_Update:Manually_Installing_a_MAR_file
 #
 
-ff() {
-	instalation_dir="/opt/firefox"
-	tmp_dir="/tmp/ff/out"
+update() {
+	instalation_dir="/opt/$app"
+	tmp_dir="/tmp/$app/out"
 	lang="en-US"
-	installed_verion=$($instalation_dir/firefox -v 2> /dev/null | grep Firefox | cut -d " " -f 3)
-	ff_url="http://download-origin.cdn.mozilla.net/pub/mozilla.org/firefox/releases/latest/update/linux-`uname -m`/$lang/"  
-	new_version=$(curl -s $ff_url | egrep -o -e "firefox\-$installed_verion\-[0-9]*\.[0-9]*\.[0-9]*\.partial\.mar" | sort -u | cut -d "-" -f 3 |cut -d "." -f 1-3)
 
+	if [ "$app" == "firefox" ]; then
+		installed_verion="$($instalation_dir/$app -v 2> /dev/null | egrep -i "$app" | cut -d ' ' -f 3)"
+	elif [ "$app" == "thunderbird" ]; then
+		installed_verion="$($instalation_dir/$app -v 2> /dev/null | egrep -i "$app" | cut -d ' ' -f 2)"
+	fi
+
+	url="http://download-origin.cdn.mozilla.net/pub/mozilla.org/$app/releases/latest/update/linux-`uname -m`/$lang/"  
+	new_version="$(curl -s $url | egrep -o -e "$app\-$installed_verion\-[0-9]*\.[0-9]*\.[0-9]*\.partial\.mar" | sort -u | cut -d '-' -f 3 |cut -d '.' -f 1-3)"
+	update_file="$(curl -s $url | egrep -o -e "$app\-$installed_verion\-[0-9]*\.[0-9]*\.[0-9]*\.partial\.mar" | sort -u)"
+	signature_file="$(curl -s $url | egrep -o -e "$app\-$installed_verion\-[0-9]*\.[0-9]*\.[0-9]*\.partial\.mar.asc" | sort -u)"
+	
 	mkdir -p $tmp_dir
 	
 	echo -en "Installed version \033[01;32m$installed_verion\033[0m, "
 
-	if [ -z $new_version ]; then
+	if [ -z "$new_version" ]; then
 		echo -e "\033[01;32mno updates!\033[0m"
 		rm -R $tmp_dir
 		exit 0
@@ -27,48 +35,47 @@ ff() {
 		echo -e "updating to the version \033[01;32m$new_version\033[0m ..."
 	fi
 	
-	wget -O /tmp/ff/out/update.mar $ff_url/$update_ff 
-	cd $instalation_dir
-	cp $instalation_dir/updater $tmp_dir
-	$tmp_dir/updater $tmp_dir $instalation_dir $instalation_dir
+	wget -q --show-progress -O $tmp_dir/update.mar $url/$update_file
+	wget -q --show-progress -O $tmp_dir/update.mar.asc $url/$signature_file
 
-	rm -R $tmp_dir
-}
+	signature_verification="$(gpg --verify $tmp_dir/update.mar.asc 2>&1)"
 
-tb() {
-	instalation_dir="/opt/thunderbird"
-	tmp_dir="/tmp/tb/out"
-	lang="en-US"
-	installed_verion=$($instalation_dir/thunderbird -v 2> /dev/null | grep Thunderbird | cut -d " " -f 2)
-	tb_url="http://download-origin.cdn.mozilla.net/pub/mozilla.org/thunderbird/releases/latest/update/linux-`uname -m`/$lang/" 
-	new_version=$(curl -s $tb_url | egrep -o -e "thunderbird\-$installed_verion\-[0-9]*\.[0-9]*\.[0-9]*\.partial\.mar" | sort -u | cut -d "-" -f 3 |cut -d "." -f 1-3)
+	no_key="$(echo $signature_verification | egrep "public key not found")"
 
-	mkdir -p $tmp_dir
-	
-	echo -en "Installed version \033[01;32m$installed_verion\033[0m , "
-
-	if [ -z $new_version ]; then
-		echo -e "\033[01;32mno updates!\033[0m"
-		rm -R $tmp_dir
-		exit 0
-	else
-		echo -e "updating to the version \033[01;32m$new_version\033[0m ..."
+	if [ ! -z "$no_key" ]; then
+		signature_key="$(echo $signature_verification | egrep "using RSA key" | cut -d ' ' -f 21)"
+		gpg --recv-keys $signature_key > /dev/null 2>&1
+		signature_verification="$(gpg --verify $tmp_dir/update.mar.asc 2>&1)"
 	fi
 
-	update_tb=$(curl -s $tb_url | egrep -o -e "thunderbird\-$installed_verion\-[0-9]*\.[0-9]*\.[0-9]*\.partial\.mar" | sort -u)
-	wget -O /tmp/tb/out/update.mar $tb_url/$update_tb
+	good_signature="$(echo $signature_verification | egrep "Good signature")"
+
+	if [ -z "$good_signature" ]; then
+		echo -e "\033[01;31mBad signature\033[0m, file not verified, aborting..."
+		exit 0
+	else
+		echo -e "\033[01;32mGood signature\033[0m, file verified!"
+	fi
+	
 	cd $instalation_dir
 	cp $instalation_dir/updater $tmp_dir
 	$tmp_dir/updater $tmp_dir $instalation_dir $instalation_dir
 
-	rm -R $tmp_dir
+	update_status="$(egrep "succeeded" $tmp_dir/update.status)"
+
+	if [ "$update_status" == "succeeded" ]; then
+		echo -e "Update complete! You can restart your $app now."
+		rm -R $tmp_dir
+	else
+		echo -e "Something went wrong, check the $tmp_dir/update.log file for more info."
+	fi
 }
 
 echo -e "\033[1mFirefox/Thunderbird updater\033[0m"
 echo -en "Usage: ff/tb for firefox and thunderbird respectively: "
 read choice
 case "$choice" in
-	"ff") ff ;;
-	"tb") tb ;;
+	"ff") app="firefox" ; update ;;
+	"tb") app="thunderbird" ; update ;;
 	*) exit 0 ;;
 esac
